+++ NEWS.md
@@ -1,3 +1,8 @@
+# clustermole (development version)
+
+* Added singscore enrichment method.
+* Added combined enrichment method.
+
 # clustermole 1.0.1
 
 * Updated documentation.